Refactored GameObject inheritance
A lot of different classes inherited from the GameObject type, which did
not really serve any purpose other than to provide an easy access to the
GameController.Instance property through the shortener "Game".

This got replaced by the GameProvider type, which can be imported via
using static. The new shorthand is "GameInstance", which can be changed
later on to something else.
It cannot be "Game", because it could be confused with
Microsoft.Xna.Framework.Game, when that namespace is imported to the
file.